+++ nominal/experimental/extractor/__init__.py
@@ -1,0 +1,23 @@
+from nominal.core.container_image import TimestampMetadata
+from nominal.experimental.extractor._extractor import (
+    Extractor,
+    ExtractorContext,
+    ExtractorError,
+    IngestType,
+    ManifestExtractorContext,
+    SingleFileExtractorContext,
+    manifest_extractor,
+    single_file_extractor,
+)
+
+__all__ = [
+    "Extractor",
+    "ExtractorContext",
+    "ExtractorError",
+    "IngestType",
+    "ManifestExtractorContext",
+    "SingleFileExtractorContext",
+    "TimestampMetadata",
+    "manifest_extractor",
+    "single_file_extractor",
+]